add searching by category

+++ frontend/components/category/category_form.jsx
@@ -2,10 +2,11 @@ var React = require('react');
 var CategoryStore = require('../../stores/category_store');
 var ApiUtil = require('../../util/api_util');
 var CategoryIndexItem = require('./category_index_item');
+var Alert = require('react-bootstrap').Alert;
 
 var CategoryForm = React.createClass({
 	getInitialState: function() {
-		return { categories: CategoryStore.all(), followedCategories: CategoryStore.followedCategories() }
+		return { categories: CategoryStore.all(), followedCategories: CategoryStore.followedCategories(), alertVisible: false }
 	},
 
 	componentDidMount: function() {
@@ -32,6 +33,22 @@ var CategoryForm = React.createClass({
 		}
 	},
 
+		showAlert: function() {
+		if (this.state.alertVisible) {
+			return (
+				<Alert className="alertMessages" bsStyle="danger" onDismiss={this.handleAlertDismiss} dismissAfter={4000}>
+					<h4>Must choose at least one category</h4>
+				</Alert>
+			)
+		} else {
+			return "";
+		}
+	},
+
+	handleAlertDismiss: function() {
+		this.setState({ alertVisible: false });
+	},
+
 	submitForm: function(e) {
 		e.preventDefault();
 		var toDelete = [];
@@ -48,13 +65,18 @@ var CategoryForm = React.createClass({
 			}
 		});
 
-		if (toCreate.length > 0) {
-			ApiUtil.createCategoryFollows(toCreate.join(", "));
+		if (toDelete.length === this.state.followedCategories.length && this.state.followedCategories.length !== 0) {
+			this.setState({ alertVisible: true });
+		} else {
+
+			if (toCreate.length > 0) {
+				ApiUtil.createCategoryFollows(toCreate.join(", "));
+			}
+			if (toDelete.length > 0) {
+				ApiUtil.deleteCategoryFollows(toDelete.join(", "));
+			}
+			this.props.modalCallback();
 		}
-		if (toDelete.length > 0) {
-			ApiUtil.deleteCategoryFollows(toDelete.join(", "));
-		}
-		this.props.modalCallback();
 	},
 
 	followedCategory: function(category) {
@@ -72,6 +94,7 @@ var CategoryForm = React.createClass({
 
 		return (
 			<form onSubmit={this.submitForm}>
+			{this.showAlert()}
 				<div className="container-fluid">
 				{
 					this.state.categories.map(function(category) {
